+++ entry/models.py
@@ -1184,7 +1184,7 @@ class Entry(ACLBase):
 
         return attr
 
-    def get_referred_objects(self, entity_name=None):
+    def get_referred_objects(self, filter_entities=[], exclude_entities=[]):
         """
         This returns objects that refer current Entry in the AttributeValue
         """
@@ -1194,10 +1194,10 @@ class Entry(ACLBase):
 
         # if entity_name param exists, add schema name to reduce filter execution time
         query = Q(pk__in=ids, is_active=True)
-        if entity_name:
-            query &= Q(schema__name=entity_name)
+        if filter_entities:
+            query &= Q(schema__name__in=filter_entities)
 
-        return Entry.objects.filter(query)
+        return Entry.objects.filter(query).exclude(schema__name__in=exclude_entities)
 
     def may_append_attr(self, attr):
         """
@@ -1811,6 +1811,7 @@ class Entry(ACLBase):
         kls,
         hint_attr_value,
         hint_entity_name=None,
+        exclude_entity_names=[],
         limit=CONFIG.MAX_LIST_ENTRIES,
         offset=0,
     ):
@@ -1823,9 +1824,12 @@ class Entry(ACLBase):
         3. Process the search results, and return. (make_search_results_for_attrv)
 
         Args:
-            hint_attr_value (str): Search string for AttributeValue
+            hint_attr_value (str): Required.
+                Search string for AttributeValue
             hint_entity_name (str): Defaults to None.
                 Search string for Entity Name
+            exclude_entity_names (list[str]): Defaults to [].
+                Entity name string list to exclude from search
             limit (int): Defaults to 100.
                 Maximum number of search results to return
             offset (int): Defaults to 0.
@@ -1851,7 +1855,9 @@ class Entry(ACLBase):
                 "ret_values": [],
             }
 
-        query = make_query_for_simple(hint_attr_value, hint_entity_name, offset)
+        query = make_query_for_simple(
+            hint_attr_value, hint_entity_name, exclude_entity_names, offset
+        )
 
         resp = execute_query(query, limit)